console/requirements.txt
openai
python-dotenv
azure-identity
jupyter
# pillow
rich